cirrus_sony: Ignore start_processing on the calibration thread
We use fp_enable_snd_device (and friends) to set up the right
routes/paths and backends before trying to use the speakers, with the
annoying caveat that start_processing is called. Usually after a reboot
this only spews `Forbidden. Calibration is in progess...` once, but if
AOSP already tries playing audio it'll get stuck in a loop with lots of
pops and clicks.

Signed-off-by: Marijn Suijten <marijns95@gmail.com>